backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-multipart
aiofiles
python-jose[cryptography]
passlib[argon2]
python-dateutil